requirements.txt
requests>=2.25.1
kafka-python>=2.0.2
python-dateutil>=2.8.2